% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{apply_kval}
\alias{apply_kval}
\title{Convert peak areas to concentration}
\usage{
apply_kval(df, standard_fnames, mw_df, standard_conc = 250,
  inj_vol = 2, standard = "13:0", soil_wt_df, vial_vol = 50)
}
\arguments{
\item{\code{vial_vol}}{}
}
\value{

}
\description{
This function converts peak areas to lipid concentration in units of nmol/g
dry soil.Peak areas should be normalized and have standards subtracted prior
to running this function.
}
\examples{


}
