% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calc_biomarkers}
\alias{calc_biomarkers}
\title{Calculate mol fractions of biomarkers and microbial groups}
\usage{
calc_biomarkers(df)
}
\arguments{
\item{\code{df}}{dataframe or tibble containing lipid concentrations. Use
output from \code{apply_kval()} function.}
}
\value{

}
\description{
This function calculates mole fractions of fungal and bacterial lipids. In
the future, this will also calculate moel fractions of Gram (+) and Gram(-)
bacteria, actinomycetes, and ___?
}
\examples{

}
