% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calc_concentration_tidy}
\alias{calc_concentration_tidy}
\title{Convert peak areas to concentration}
\usage{
calc_concentration_tidy(
  df,
  standard_fnames,
  mw_df = lipid_reference,
  standard_conc = 250,
  inj_vol = 2,
  standard = "13:0",
  soil_wt_df,
  vial_vol = 20
)
}
\arguments{
\item{\code{vial_vol}}{}
}
\value{

}
\description{
This function converts peak areas to lipid concentration in units of nmol/g
dry soil.Peak areas should be normalized and have standards subtracted prior
to running this function.
}
\examples{


}
