% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_base.R
\name{calculate_indicators}
\alias{calculate_indicators}
\title{Calculate concentration and relative abundance of indicator lipids}
\usage{
calculate_indicators(df, soil_wt_df)
}
\arguments{
\item{df}{Dataframe or tibble generated from \code{process_peak_area()} or
a dataframe/tibble with the same fields and format}

\item{soil_wt_df}{Dataframe or tibble containing the dry weight (g) of soil
used for lipid extraction for each sample.}

\item{dat}{Dataframe or tibble with named PLFA peak data following format
specified by \code{import_batch()}.}
}
\value{
Tibble similar to input with the addition of columns for indicator
groups (including total biomass), relative abundance, concentration, and
fungal:bacterial lipid ratio.
}
\description{
This function aggregates lipid concentration data into indicator groups
(eg. fungal lipids, bacterial lipids) and calculates their relative
abundance. Indicator groups reported as the sum of lipid concentration/
abundance of all lipids in that group. This also calculates total biomass
and fungal:bacterial lipid ratio.
}
\examples{

}
