% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_calcs_base.R
\name{correct_iso_base}
\alias{correct_iso_base}
\title{Process del13C values}
\usage{
correct_iso_base(df, d13c_correction, methanol_13c, min_height = NULL)
}
\arguments{
\item{\code{df}}{peak list dataframe or tibble. This should contain the
following columns at a minimum: Batch, DataFileName, RetTimeSecs,
MajorHeightnA, TotalPeakArea1, DisplayDelta1, Name. This can be the output
from import_batch() or import_batch_multi().}

\item{\code{c13_correction}}{difference between measured USGS40 value and
expected USGS40 value (-26.39 per mil). If IonVantage output has already
been corrected, then enter 0.}

\item{\code{methanol_13c}}{13C signature of the methanol used for
methylation.}
}
\value{
Returns a tibble with columns from input dataframe and a corrected
d13C column.
}
\description{
This function corrects the raw del13C output for methylation and normalizes
to the del13C standard USGS40 (L-glutamic acid).
}
\examples{

}
