% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_calcs_base.R
\name{correct_iso_base2}
\alias{correct_iso_base2}
\title{Process del13C values}
\usage{
correct_iso_base2(df, d13c_correction, meth_13c_df, min_height = NULL)
}
\arguments{
\item{\code{df}}{peak list dataframe or tibble. This should contain the
following columns at a minimum: Batch, DataFileName, RetTimeSecs,
MajorHeightnA, TotalPeakArea1, DisplayDelta1, Name. This can be the output
from import_batch() or import_batch_multi().}

\item{\code{d13c_correction}}{difference between measured USGS40 value and
expected USGS40 value (-26.39 per mil). If IonVantage output has already
been corrected, then enter 0.}

\item{\code{meth_13c_df}}{Dataframe identifying the 13C signature of the
methanol used for each batch. This dataframe should have 2 columns: Batch
and Meth13C.}
}
\value{
Returns a tibble with columns from input dataframe and a corrected
d13C column.
}
\description{
This function corrects the raw del13C output for methylation and normalizes
to the del13C standard USGS40 (L-glutamic acid). This function is an update
of the correct_iso_base() function that uses a dataframe matching methanol
lot to batch so that we can use different methanol lots for an experiment.
}
\examples{

}
