% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs_base.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Identify duplicate peak names within samples}
\usage{
find_duplicates(df)
}
\arguments{
\item{df}{Dataframe or tibble containing the following columns: Batch,
DataFileName, RetTimeSecs, MajorHeightnA, TotalPeakArea1, DisplayDelta1,
Name.}
}
\value{
Dataframe indicating the sample containing duplicate lipids, and the
lipids within the sample that were duplicated.
}
\description{
\code{find_duplicates} Identify samples with naming errors resulting in a
lipid name being given to multiple peaks in a sample.
}
\examples{

}
