% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs_base.R
\name{find_missing}
\alias{find_missing}
\title{\code{find_missing} Identifies samples that are missing lipids that are
expected to be in all samples. For example, 13:0 and 19:0 lipids are added
to all samples as standard and surrogate standard, so these lipids missing
may indicate issues with the run or with the naming. For peat samples, 16:0
is expected to appear in all samples, as this lipid is present in Sphagnum.}
\usage{
find_missing(df, lipids = c("13:0", "16:0", "19:0"))
}
\arguments{
\item{df}{Dataframe or tibble containing the following columns: Batch,
DataFileName, RetTimeSecs, MajorHeightnA, TotalPeakArea1, DisplayDelta1,
Name.}

\item{lipids}{Character vector listing the lipids that should be found in
all samples.}
}
\value{
Dataframe indicating the sample containing missing the specified
lipids, and the lipids within the sample that were missing.
}
\description{
\code{find_missing} Identifies samples that are missing lipids that are
expected to be in all samples. For example, 13:0 and 19:0 lipids are added
to all samples as standard and surrogate standard, so these lipids missing
may indicate issues with the run or with the naming. For peat samples, 16:0
is expected to appear in all samples, as this lipid is present in Sphagnum.
}
\examples{

}
