% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_base.R
\name{import_batch_multi}
\alias{import_batch_multi}
\title{Import multiple PLFA peak lists (Excel files) in a directory}
\usage{
import_batch_multi(dir, keyword)
}
\arguments{
\item{dir}{Full path to the directory containing the PLFA data Excel files
to be imported. The input Excel files must contain Batch, DataFileName,
RetTimeSecs, MajorHeightnA, TotalPeakArea1, DisplayDelta1, and Name columns.}
}
\value{
A named list. Each element is a tibble suitable for downstream
analyses in plfaR. Tibbles contain the input columns as well as
BatchDataFileName.
}
\description{
This is the first function in the plfaR workflow, loading named PLFA peak
list from mutliple batches for use in downstream analysis.
}
\examples{

}
