% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_calcs_base.R
\name{indic_iso_base}
\alias{indic_iso_base}
\title{Calculate d13C for indicator groups}
\usage{
indic_iso_base(df, sample_metadata)
}
\arguments{
\item{\code{df}}{dataframe containing corrected d13C values for individual
lipids. Normally, this would be the output of correct_iso_base().
This should contain the following columns at a minimum: Batch, DataFileName,
BatchDataFileName, d13C_corrected, Name.}

\item{\code{sample}{}
}
\value{
Returns a tibble with corrected d13C values for each indicator
group.
}
\description{
This function aggregates individual lipid d13C signatures by indicator
group. Current implementation uses a straight average. Future versions
may offer the option to select between average and weighted average.
}
\examples{

}
