% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_reference}
\alias{lipid_reference}
\title{Molecular weights of microbially-derived phospholipids}
\format{

}
\usage{
lipid_reference
}
\description{
COmpiled from Jess Gutknecht and literature review. Weights are for
methylated lipids and are rounded to the nearest whole number.
}
\keyword{datasets}
