% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_batch}
\alias{load_batch}
\title{Load single batch PLFA peak list}
\usage{
load_batch(file_path)
}
\arguments{
\item{file_path}{Full file path, including file extension for the Excel file
containing PLFA data. The input Excel file must contain Batch, DataFileName,
RetTimeSecs, MajorHeightnA, TotalPeakArea1, DisplayDelta1, and Name columns.}
}
\value{
Tibble suitable for downstream analyses in plfayer. Contains the
input columns as well as BatchDataFileName.
}
\description{
This is the first function in the plfayer workflow, loading named PLFA peak
list from a single batch for use in downstream analysis.
}
\examples{

}
