% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{normalize_area}
\alias{normalize_area}
\title{Normalize peak areas across batches}
\usage{
normalize_area(df)
}
\arguments{
\item{\code{df}}{peak list as dataframe or tibble, loaded using
\code{load_batch()} function.}
}
\value{
Returns with same fields as input dataframe as well as
Normalized CFactor (conversion factor; used for normalization),
NormalizedArea and other intermeidates used in normalization
calculations.
}
\description{
GC response varies varies across batches, producing different peak areas
for same lipid cocncentration. This function normalizes the peak area
across batches for consistent results within an analysis. Current
implementation normalizes to the batch with the smallest mean peak area
(non-detects not treated as 0 and not included in the average). Averages
used for the correction factor only includes experimental samples, blanks
(B), and methylation controls (C) (ie. internal standard, M1M2, and FAME).
This is consistent with the Excel example calculations provided by Jess.
}
\examples{


}
