% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_base.R
\name{process_peak_area}
\alias{process_peak_area}
\title{Convert peak areas to concentration (nmol/g dry soil)}
\usage{
process_peak_area(
  dat,
  standard_fnames,
  mw_df = lipid_reference,
  standard_conc = 250,
  inj_vol = 2,
  standard = "13:0",
  soil_wt_df,
  vial_vol = 50,
  blanks
)
}
\arguments{
\item{dat}{Dataframe or tibble with named PLFA peak data following format
specified by \code{import_batch()}.}

\item{standard_fnames}{Character vector with names of samples to be used as
standards for calculating the kval. Names should match names in DataFileName
column of input dataframe.}

\item{mw_df}{Dataframe including all lipids (fame column) of interest and
their associated molecular weights (molecular_weight_g_per_mol column). By
default, this function uses the included \code{lipid_reference} dataframe.}

\item{standard_conc}{Numeric value indicating the concentraion of standard
in ng/uL. Default is 250 ng/uL.}

\item{inj_vol}{Volume of sample injected into GC in uL. Default is 2uL.}

\item{standard}{String indicating the lipid used as the standard. Default is
13:0.}

\item{soil_wt_df}{Dataframe or tibble containing the dry weight (g) of soil
used for lipid extraction for each sample.}

\item{vial_vol}{Volume solution run on GC. Default is 50 uL}

\item{blanks}{Vector of the names of samples run as blanks. This should
include all samples you want to use for peak subtraction (eg. 13:0 and 19:0)
in most cases.}
}
\value{
Tibble similar to input with the addition of concentration column.
}
\description{
This function takes a formatted peak list and carries out blank subtraction
and calculates kval in order to convert peak area for each peak to
concentration.
}
\examples{

}
