% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{quality_check}
\alias{quality_check}
\title{Inspect PLFA peak list quality.}
\usage{
quality_check(df)
}
\arguments{
\item{df}{Dataframe or tibble containing the following columns: Batch,
DataFileName, RetTimeSecs, MajorHeightnA, TotalPeakArea1, DisplayDelta1,
Name.}
}
\value{
List of dataframes indicating samples with duplicate peak names,
samples missing standard peaks, and the distribution of lipids among
samples.
}
\description{
\code{quality_check} returns a list of dataframes containing the following
information related to the quality of the data in the peak list:
Samples with duplicate peak names
Samples that are missing one or more of the standards peaks 13:0, 16:0, 19:0
The distribution of lipids among samples
}
\examples{


}
