% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_base.R
\name{subtract_blanks}
\alias{subtract_blanks}
\title{Subtract peak area from control and standard lipids}
\usage{
subtract_blanks(df, blanks)
}
\arguments{
\item{df}{Dataframe or tibble with named PLFA peak data following format
specified by \code{import_batch()}.}

\item{blanks}{Vector of the names of samples run as blanks. This should
include all samples you want to use for peak subtraction (eg. 13:0 and 19:0)
in most cases.}
}
\value{
Tibble similar to input with the addition of AreaMinusBlanks column
showing peak areas minus area contributed by standard/control lipids.
}
\description{
All samples are typically run with a 19:0 extraction control and 13:0
standard. This function subtracts peak area contributed by these additions
prior to converting peak area to concentration.
}
\examples{

}
