% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{subtract_blanks_tidy}
\alias{subtract_blanks_tidy}
\title{Subtract surrogate and methylation standard peak areas}
\usage{
subtract_blanks_tidy(
  df,
  blanks = c("100.raw", "101.raw"),
  lipids = c("13:0", "19:0")
)
}
\arguments{
\item{\code{df}}{Dataframe or tibble with NormalizedPeakArea. Use the output
from \code{normalize_area_tidy()} function.}

\item{\code{blanks}}{}

\item{\code{lipids}}{}
}
\value{

}
\description{
Gutknecht lab PLFA runs all include a 19;0 surrogate standard spiked
into soil samples before extraction, and extracted lipids are suspended
in 13:0 control solution. We need to subtract the peak area due to these
samples, so that we only measure lipids from the soil sample. This function
subtracts 13:0 and 19:0 peak areas to account for this. User must specify
the DataFileNames associated with the controls used for peak subtraction.
User may also select peaks for subtraction (default is 13:0 and 19:0).
}
\details{
Might want to integrate id of blanks etc. in the metadata.
}
\examples{


}
