% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{subtract_blanks}
\alias{subtract_blanks}
\title{Subtract surrogate and methylation standard peak areas}
\usage{
subtract_blanks(df, biomarkers = c("13:0", "19:0"))
}
\arguments{
\item{\code{df}}{Dataframe or tibble with NormalizedPeakArea. Use the output
from \code{normalize_area()} function.}

\item{\code{biomarkers}}{}
}
\value{

}
\description{
Gutknecht lab PLFA runs all include a 19;0 surrogate standard spiked
into soil samples before extraction, and extracted lipids are suspended
in 13:0 control solution. We need to subtract the peak area due to these
samples, so that we only measure lipids from the soil sample. This function
subtracts corrects 13:0 and 19:0 peak areas to account for this. The current
implementation uses the average peak area of all samples from the batch,
excluding internal standards, M1M2, and FAME. These averages do include 0s.
This implementation is consistent with the example Excel calcs provided by
Jess. Editorial: I think it would make the most sense to subtract only the
peak areas from blanks and methylation standards (B and C). I think this is
also what the protocol states.
}
\examples{


}
